+++ frontend/src/components/Ticket/index.js
@@ -262,6 +262,7 @@ const Ticket = () => {
           ticketId={ticket.id}
           isGroup={ticket.isGroup}
           markAsRead={!isObserving}
+          readOnly={isObserving}
         ></MessagesList>
         <MessageInput
           ref={messageInputRef}